requirements.txt
amqp==5.2.0
asgiref==3.8.0
billiard==4.2.0
celery==5.3.6
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
iniconfig==2.0.0
kombu==5.3.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
